Split tests for modularity, and shorter test names
